pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry]
authors = ["Hiroshi Tsuyuki <kagemeka1@gmail.com>"]
classifiers = []
description = "Anikore API for Python"
documentation = "https://anikore.readthedocs.io"
exclude = [
  "./**/tests/",
  "./**/*_test.py",
  "./**/test_*.py",
  "./**/_test_*.py",
]
homepage = "https://kagemeka.github.io/anikore"
include = []
keywords = ["anikore", "api"]
license = "MIT"
maintainers = ["Hiroshi Tsuyuki <kagemeka1@gmail.com>"]
name = "anikore"
packages = [
  {include = "anikore", from = "src"},
]
readme = "README.md"
repository = "https://github.com/kagemeka/anikore"
version = "0.2.0"

[tool.poetry.dependencies]
PyMySQL = "^"
PyYAML = "^"
SQLAlchemy = "^"
Scrapy = "^"
aiodns = "^"
aiohttp = "^"
beautifulsoup4 = "^"
boto3 = "^"
cchardet = "^"
cryptography = "^"
filesystem-python = "0.2.0"
lxml = "^"
optext-python = "0.1.1"
pandas = "^"
python = ">=3.8,<3.11"
requests = "^"
selenium = "^"
tqdm = "^"

furo = {version = "^", optional = true}
myst-parser = {version = "^", optional = true}
pdoc3 = {version = "^", optional = true}
pydata-sphinx-theme = {version = "^", optional = true}
python-docs-theme = {version = "^", optional = true}
sphinx = {version = "^", optional = true}
sphinx-book-theme = {version = "^", optional = true}
sphinx-theme-pd = {version = "^", optional = true}
sphinx_rtd_theme = {version = "<=2.0.0", optional = true}
sphinxcontrib-mermaid = {version = "^", optional = true}

[tool.poetry.extras]
docs = [
  "sphinx",
  "furo",
  "myst-parser",
  "sphinx_rtd_theme",
  "sphinxcontrib-mermaid",
  "pdoc3",
  "python-docs-theme",
  "sphinx-theme-pd",
  "sphinx-book-theme",
  "pydata-sphinx-theme",
]

[tool.poetry.dev-dependencies]
autopep8 = "^"
bandit = "^"
black = "^"
blacken-docs = "^"
flake8 = "^"
isort = "^"
mypy = "^"
pre-commit = "^"
pydocstyle = "^"
pylint = "^"
pyproject-flake8 = "^"
pytest = "^"
pytest-asyncio = "^"
types-PyYAML = "^"
types-requests = "^"
types-setuptools = "^"

pycodestyle = "^2.8.0"
# FORMATTING ------------------------------------------------------------------------------
[tool.black]
# github: https://github.com/psf/black
# docs: https://black.readthedocs.io/en/stable/
# compatible configurations: https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html
line-length = 79
target-version = ["py38", "py39", "py310"]
verbose = true

[tool.autopep8]
# github: https://github.com/hhatto/autopep8
aggressive = 3
ignore = "E501,W6" # or ["E501", "W6"]
in-place = true
max_line_length = 79
recursive = true

[tool.isort]
# github: https://github.com/PyCQA/isort
# docs: https://pycqa.github.io/isort/
# config files -> https://pycqa.github.io/isort/docs/configuration/config_files.html
# options: https://pycqa.github.io/isort/docs/configuration/options.html
profile = "black"
# profiles: https://pycqa.github.io/isort/docs/configuration/profiles.html
line_length = 79

# LINTING ------------------------------------------------------------------------------
[tool.flake8]
# github: https://github.com/PyCQA/flake8
# docs: https://flake8.pycqa.org/en/latest/
# options: https://flake8.pycqa.org/en/latest/user/options.html
count = 'true'
exclude = ["docs/", "scripts/"]
exit_zero = 'true'
ignore = 'E203'
max_complexity = 10 # https://en.wikipedia.org/wiki/Cyclomatic_complexity
max_line_length = 79
show_source = 'true'
statistics = 'true'
verbose = 'true'

# TYPE CHECKING ------------------------------------------------------------------------------
[tool.mypy]
# github: https://github.com/python/mypy
# docs: https://mypy.readthedocs.io/en/stable/
# options: https://mypy.readthedocs.io/en/stable/config_file.html
allow_redefinition = false
allow_untyped_globals = false
cache_dir = ".mypy_cache"
cache_fine_grained = false
check_untyped_defs = true
color_output = true
disallow_any_decorated = true
disallow_any_explicit = true
disallow_any_expr = false
disallow_any_generics = true
disallow_any_unimported = false
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = false
disallow_untyped_decorators = true
disallow_untyped_defs = true
error_summary = true
exclude = ["docs/"]
follow_imports = 'normal'
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
incremental = true
local_partial_types = true
no_implicit_optional = true
no_implicit_reexport = true
no_strict_optional = false
pdb = true
pretty = true
python_version = "3.10"
raise_exceptions = true
scripts_are_modules = true
show_absolute_path = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
show_none_errors = true
show_traceback = true
skip_cache_mtime_checks = false
skip_version_check = false
sqlite_cache = false
strict = true
strict_equality = true
strict_optional = true
verbosity = 0
warn_incomplete_stub = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = false
warn_unreachable = true
warn_unused_configs = false
warn_unused_ignores = true

[tool.pyre]
# github: https://github.com/facebook/pyre-check
# docs: https://pyre-check.org/

# TESTING ------------------------------------------------------------------------------
[tool.pytest.ini_options]
# github: https://github.com/pytest-dev/pytest
# docs: https://docs.pytest.org
cache_dir = "/tmp/.pytest_cache"

[tool.tox]
# github: https://github.com/tox-dev/tox
# docs: https://tox.wiki/en/latest/
